Implemented Drive UI and some functionality
Implemented Drive UI. Contains a Seekbar for power controls, and
directional arrow ImageButtons to control where the robot goes in a
single direction. Started to implement button controls using
OnTouchListener so when the buttons are held, the robot should move, but
when released, it should stop moving. Could not test full functionality
because bluetooth connection is not implemented. - Kyle