Lots of updates
Added permissions for Bluetooth. Added assets for Bluetooth icons
(probably will remove). Implemented a DialogFragment that displays
Bluetooth devices already registered to the phone. Bluetooth elements
not fully implemented. Implemented a ListAdapter for the list in the
DialogFragment. Created the DialogFragment XML as well as Cells for the
List in the DialogFragment.  Modified ConnectionFragment and updated
it's XML to have a ToggleButton instead of a regular Button (more
efficient). Added the BroadcastReceiver to the ConnectionFragment. Added
onActivityCreated to both fragments. Modified onCreateView in both
fragments. Did some refractoring for naming. Added some color and string
resources. Still working on this, but figured it would be good to commit
my progress. Feel free to look at it @Beals1194 @andrewsbryanc
@bwurtsmith
-Kyle