Implemented BluetoothConnection
The app can now pair with a device. Manifest was updated to include the
BTDialog activity.  Refactored BTConnectFragment to BTConnection.
BTDialog is implemented as a intent and runs on it's own screen.
ConnectionFragment contains the BluetoothSocket and the BluetoothDevice
we're connected to. ToggleButton is fixed on screen to handle events
properly.
-Kyle