Bluetooth Connection works
Added setData method to BTDeviceListAdapter
Got rid of redundant unused imports in BTDialog and implemented
BroadcastReceiver in BTDialog. Added ProgressDialog to BTDialog so we
can see it's actually scanning for devices. Added onDestroy to BTDialog
so we unregister BroadcastReceiver.
Got rid of unused BroadcastReceiver in ConnectionFragment. Added
BluetoothAdapter to ConnectionFragment. Everyone is here, no need for
code review
--
Kyle