Updated small things, added toolbar to top of app
So small things I changed.
Battery now displays as a percentage.
Based on the connection state, it will change what's on screen. Feels
very nice, and looks very nice.
Implemented an ActionBar so we can use menus.
I began creating a menu so @bwurtsmith can begin working on it.
I eliminated the big label on the connection fragment we had before
since the ActionBar and Tabs were covering it up. It was really there to
fill space, so now we don't need it. I also got rid of the string
resource for it.
Added menu_main so we can have preference menus and stuff. I temporarily
threw in a second menu option called Drive Mode. We could switch between
drive modes here? Like use the directional by default and then switch to
tilt movement.
Added resources for menu and ActionBar